Added basic Jupyter tests and fix issues
