+++ glue_vispy_viewers/common/qt/data_viewer.py
@@ -18,9 +18,18 @@ BROKEN_PYQT5_MESSAGE = ("The version of PyQt5 you are using does not appear to "
 class BaseVispyViewer(BaseVispyViewerMixin, DataViewer):
 
     _options_cls = VispyOptionsWidget
+    tools = BaseVispyViewerMixin.tools
     subtools = {'save': ['vispy:save']}
 
     _toolbar_cls = VispyQtToolbar
+
+    # If imageio is available, we can add the record icon
+    try:
+        import imageio  # noqa
+    except ImportError:
+        pass
+    else:
+        tools.insert(1, 'vispy:record')
 
     def __init__(self, session, state=None, parent=None):
         super().__init__(session, state=state, parent=parent)
