+++ glue_vispy_viewers/volume/jupyter/volume_viewer.py
@@ -15,7 +15,7 @@ class JupyterVispyVolumeViewer(VispyVolumeViewerMixin, IPyWidgetView):
 
     def __init__(self, *args, **kwargs):
         # Vispy and jupyter_rfb don't work correctly on Linux unless DISPLAY is set
-        if 'DISPLAY' not in os:
+        if 'DISPLAY' not in os.environ:
             os.environ['DISPLAY'] = ':0'
         super().__init__(*args, **kwargs)
         self.setup_widget_and_callbacks()
