+++ setup.cfg
@@ -42,6 +42,8 @@ pyside =
     qtpy
     glue-qt>=0.1.0
     PySide6
+jupyter =
+    glue-jupyter
 all =
     imageio
 
@@ -50,6 +52,8 @@ glue_vispy_viewers.common = *.png
 glue_vispy_viewers.common.qt = *.ui
 glue_vispy_viewers.scatter.qt = *.ui
 glue_vispy_viewers.volume.qt = *.ui
+glue_vispy_viewers.scatter.jupyter = *.vue
+glue_vispy_viewers.volume.jupyter = *.vue
 glue_vispy_viewers.qt.tests = data/*.glu
 glue_vispy_viewers.extern = README.md, VISPY_LICENSE
 glue_vispy_viewers.extern.vispy = io/_data/*.npy, html/static/js/*.js, app/tests/*.ui
