+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist =
-    py{38,39,310,311}-{test}-{dev}{,pyqt63,pyside63,pyqt64,pyside64}
+    py{38,39,310,311}-{test}-{dev}{,pyqt63,pyside63,pyqt64,pyside64,jupyter}
     codestyle
 requires = pip >= 18.0
            setuptools >= 30.3.0
@@ -22,6 +22,7 @@ extras =
     test: test
     pyqt63,pyqt64: pyqt
     pyside63,pyside64: pyside
+    jupyter: jupyter
 commands =
     test: pip freeze
     test: pytest --pyargs glue_vispy_viewers --cov glue_vispy_viewers {posargs}
